+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.11.0-dev
 
 * Enhancements
+  * Add `Plug.Conn.async_assign/3` and `Plug.Conn.await_assign/3` to start and await for assigns asynchronously, mimic'ing `Task.async/1` and `Task.await/2` behaviour
   * Add `Plug.Conn.WrapperError` to propagate an error with the connection for better debugging by either `Plug.Debugger` or `Plug.ErrorHandler`
 
 * Bug fixes